swc-installation-test-2.py: Bump minimum IPython version to 1.0
I see the README recommends IPython >= 1.0 [1], but this script was
only looking for >= 0.13.  The busy bees at IPython have recently cut
2.3.0 (2014-10-01 [2]).  IPython 2.x has a bunch of changes including
deprecating the file/ prefix for local files [3] and adding a modal
interface [4].  I asked for feedback on what versions we were
supporting, and got this from Matt [5]:

On Thu, Sep 18, 2014 at 11:40:04AM -0700, Matt Davis wrote:
> I just did a quick check of opening a 2.2 made notebook with IPython
> 1.1 and it worked fine. There were some new features in IPython 2
> like notebook signing and the idea of "trusted" notebooks, but the
> format seems to be compatible with IPython 1.

So this commit bumps the minimum version to 1.0.  We'll revisit this
compatibility after the IPython folks cut 3.0.

[1]: https://github.com/swcarpentry/bc/blob/v2014.06/README.md#faq
[2]: https://pypi.python.org/pypi/ipython/2.3.0
[3]: http://ipython.org/ipython-doc/dev/whatsnew/version2.0.html#directory-navigation
[4]: http://ipython.org/ipython-doc/dev/whatsnew/version2.0.html#modal-user-interface
[5]: https://github.com/swcarpentry/bc/issues/724#issuecomment-56083316