Change manual setup instructions
The old instructions would fail on older versions of Git.
No one has complained about the Git commands in the create
script yet, though, so this updates the manual instructions
to more or less follow the exact same steps done by the
create script, including the same Git commands.

Now, instead of creating and cloning an empty repo, users
are instructed to create a new repo, but clone the workshop-template
repo and add their new repo as a remote.