Include explicit `git push --set-upstream` instruction
When following the instructions, I tried to run `git push` to push my changes, but the local `gh-pages` branch was still tracking the upstream branch, rather the branch in the `origin` remote repository where I wanted it to go. This adds an explicit example to make sure you try to push to your own repo.