Add information to preview Eventbrite widget
We should avoid mixing HTTP and HTTPS. For accomplish this we
are using `//` at the start of the links at the homepage so that
the same protocol is used. For that reason to preview/test the
links you need to launch a HTTP server.

Thanks to @aaren for reviews and comments.