Improving checking of country names.
The country name in the YAML header is used to match the filename of a flag icon.
It must be properly capitalized and use hyphens instead of spaces --- abbreviations
like "USA" and spaced names like "United Kingdom" don't work.  Rather than trying
to encode those rules, this change relies on a table of acceptable names.