README.md: Change headers ("Students" -> "For leaners")
This is based on swcarpentry/workshop-template@5451fb9c1 (Enumerating
setup/run instructions for installation test scripts, 2014-12-17).
Greg suggested he switch from "Students" to "Learners" because some
professors attend workshops as students but don't feel comfortable
being called students [1]?  Or something like that.  I don't mind
either way, so we'll just go with Greg's preference.

The addition of a "For" in the section headings seems like a good
change though, since it makes it clear that the noun in the heading
marks the intended audience for that section, not the intended
subject.

I went with 'Sentence case' titles because I think they're more
readable and it's easier to be consistent.  There are a number of
folks who agree with me [2,3], including (apparently) the Oxford Guide to
Style [4] (I'm ok with title case for book and organization titles ;).
Although the American Psychological Association perfers 'Title Case'
for the first two levels of headings as well [5].

[1]: https://github.com/swcarpentry/workshop-template/pull/89#issuecomment-67718126
[2]: http://www.snappysentences.com/sentence-case-v-title-case/
[3]: http://www.stickycontent.com/blog/are-you-team-title-case-or-team-sentence-case.php
[4]: http://ux.stackexchange.com/a/48620
[5]: http://blog.apastyle.org/apastyle/2012/03/title-case-and-sentence-case-capitalization-in-apa-style.html

Based-on-patch-by: Greg Wilson <gvwilson@third-bit.com>