Add words about the need for a web browser
Neither IPython notebooks nor github.com will
work well with old web browsers. This commit
adds text to the start of the setup section
asking participants to make sure they have an
up to date web browser (linking to the list
of supported browsers for github and the
IPython notebook). We also sugest that the
OS should be updated.

This fixes #104, but the exact wording may
need a tweak.