Update web browser instructions.
This moves the requierments into the relevent
"section" (git and python), and lists the supported
versions. This way e.g. the Python requierment will
go away for an R workshop.

I've left a generic "you will need a web browser" at
the top of the insstallation instructions. People will
need some sort of browser to access an Etherpad and any
documentation.

Links to the requierments page are in comments (so that
these can easily be checked and updated in fututre).