Remove version numbers from Python web browsers
Rather than list the exact version numbers of
"supported" browsers needed to run IPython notebooks,
just say that the current versions are supported
(which is true) and that some older versions are not
(which is also true). The actual constraint (for everything
other than IE) will come from GitHub anyway (which only
supports "current versions" of these browsers, but
supports IE 9).