README.md: Shorten the link text for the GitHub importer
To avoid distracting folks with a large, in-your-face URL.  They can
always mouse over the link (or just click it) if they want to see
where it points.

I've also used a reference-style link.  The text we use for the link
and the URL we're pointing at are pretty orthogonal, so splitting them
up makes it easier to version each independently with a line-based
diff/merge tool like Git.  It's also easier to read the source text
without the long URL cluttering up the human-readable portion.