Add instructions to select Windows' console when installing Git for Windows
Git for Windows now (at least for 2.5+) defaults to using MinTTY as the terminal
emulator. This causes nano and other programs to fail to work properly and we
have been unable to come up with consistent ways to fix this. This changes the
setup instructions to have students change the default to use the Windows' default
console, which works fine with our stack.