index.html: Comment out setup/ link to make installation-tests opt-in
Sometimes instructors forget to customize their CHECKS, but learners
still follow this link and run the stock checks, resulting in a bunch
of confusing errors about missing packages that the learner doesn't
actually need [1,2,3,4].  Instead of expecting instructors to edit
CHECKS or remove the setup/ reference from index.html (a default that
is right for nobody), this commit comments that setup/ reference out
(a default that is right for folks who don't want
installation-checks).  Instructors who *do* want installation checks
now need to do two things (uncomment the setup/ reference and update
CHECKS), but with both steps listed in CUSTOMIZATION they are more
likely to get that right.

[1]: https://github.com/swcarpentry/workshop-template/issues/136
[2]: https://github.com/swcarpentry/workshop-template/issues/180
[3]: https://github.com/swcarpentry/workshop-template/issues/181
[4]: https://github.com/swcarpentry/workshop-template/issues/258