tools/update-downloads.py: Add auto-update tooling for downloads
It's easier if lessons declare these themselves, and we just list the
lessons we're using.  This script fetches lesson-defined download
lists and updates the Downloads section appropriately.

I'd rather have used an HTML parser to make the substitution, because
line.startswith(...) is brittle.  Unfortunately, index.html is part
YAML, part HTML, and part Jekyll, so there's no out-of-the-box parser.

The lessons.json config format is compatible with my recommended
dependency format [1].

[1]: https://github.com/wking/swc-setup-installation-test/pull/9