Updated SQL install instructions for swc and dc
SWC changes: wrote a bash script that is tested against git-bash and the linux subsystem for windows thingo (via someone else in slack) to pull a copy of sqlite from sqlite.org and add it to local bin and then add that local bin to bash path.

DC changes: there's a pull request asking for the sqlitebrowser, so I added a check if the workshop is dc, which just points them at the sqlite browser download page.