Update incorrect Bash instructions and clarify
Either 3 items need to be selected, in which case the verbiage “Leave all three items selected” is wrong, as only 2 items are selected by default. Or all three items would need to be selected, requiring to change the verbiage to “Ensure that all three items are selected and click on “Next”.” I think however, that only the two items are required as that’s what worked in the past.